Move snapshot_test function out of reporter
